requirements.txt
anyio==3.7.1
black==23.3.0
click==8.1.3
fastapi==0.99.1
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pydantic==1.10.11
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
